ml/requirements.txt
pymongo>=4.0
xgboost>=2.0
scikit-learn>=1.3
python-dotenv>=1.0
fastapi>=0.100
uvicorn>=0.22
numpy>=1.24
